requirements.txt
Flask==1.1.2
Flask-Compress==1.8.0
Flask-Migrate==2.1.1
Flask-MySQLdb==0.2.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.0
gunicorn==20.1.0
gevent==21.12.0
jsonschema==3.2.0
mysqlclient==1.4.6
requests==2.25.0
SQLAlchemy==1.4.14
python-dateutil==2.8.1
PyJWT==2.3.0
flask-restx==0.5.1
flask-jwt-extended==4.3.1